update
update to adopt new versions of C++ and python

+++ pyunidoe/pyunidoe.py
@@ -172,7 +172,7 @@ def design_update(n, s, q, x, crit="CD2"):
         raise ValueError("Invalid design table.")
     if (not isinstance(x, np.ndarray)):
         raise ValueError("The design matrix must be a numpy array.")
-    elif ((np.min(x) <= 0) | (x.dtype != np.int)):
+    elif ((np.min(x) <= 0) | (x.dtype != int)):
         raise ValueError("The values in design matrix should be integers: 1,2,3,...")
 
     success_flag = False
@@ -309,7 +309,7 @@ def gen_ud(n, s, q, init="rand", initX=np.array([[]]), crit="CD2", maxiter=100, 
             raise ValueError("initX must be a numpy array.")
         elif ((n != initX.shape[0]) | (s != initX.shape[1])):
             raise ValueError("The size of the input design matrix does not match the given n,s.")
-        elif ((1 != (np.min(initX))) | (q != (np.max(initX))) | (initX.dtype != np.int)):
+        elif ((1 != (np.min(initX))) | (q != (np.max(initX))) | (initX.dtype != int)):
             raise ValueError("The values of the input design matrix should be integers within: 1,2,3...,q.")
         elif (any(np.array([v for i in range(initX.shape[1]) for v in np.unique(initX[:, i], return_counts=True)[1]]) > n / q)):
             raise ValueError("initX does not follow a balanced design.")
@@ -422,7 +422,7 @@ def gen_aud(xp, n, s, q, init="rand", initX=np.array([[]]), crit="CD2", maxiter=
         raise ValueError("xp must be a numpy array.")
     elif ((n <= xp.shape[0]) | (s != xp.shape[1])):
         raise ValueError("The size of the existing design matrix xp does not match the given n,s.")
-    elif ((xp.dtype != np.int) | (1 > np.min(xp)) | (q < np.max(xp))):
+    elif ((xp.dtype != int) | (1 > np.min(xp)) | (q < np.max(xp))):
         raise ValueError("The values of the existing design matrix x0 should be integers within: 1,2,3...,q.")
     elif (any(np.array([v for i in range(xp.shape[1]) for v in np.unique(xp[:, i], return_counts=True)[1]]) > n / q)):
         raise ValueError("xp does not follow a balanced design, please increase the number of n or remove duplicated elements (per column) in xp.")
@@ -431,7 +431,7 @@ def gen_aud(xp, n, s, q, init="rand", initX=np.array([[]]), crit="CD2", maxiter=
             raise ValueError("initX must be a numpy array.")
         elif ((n <= initX.shape[0]) | (s != initX.shape[1])):
             raise ValueError("The size of the input design matrix does not match the given n,s.")
-        elif ((np.min(initX) < 1) | (q < np.max(initX)) | (initX.dtype != np.int)):
+        elif ((np.min(initX) < 1) | (q < np.max(initX)) | (initX.dtype != int)):
             raise ValueError("The values of the input design matrix should be integers within: 1,2,3...,q.")
         elif (any(np.array([v for i in range(initX.shape[1]) for v in np.unique(initX[:, i], return_counts=True)[1]]) > n / q)):
             raise ValueError("initX does not follow a balanced design.")
@@ -544,7 +544,7 @@ def gen_aud_col(xp, n, s, q, init="rand", initX=np.array([[]]), crit="CD2", maxi
         raise ValueError("xp must be a numpy array.")
     elif ((n != xp.shape[0]) | (s <= xp.shape[1])):
         raise ValueError("The size of the existing design matrix xp does not match the given n,s.")
-    elif ((xp.dtype != np.int) | (1 != np.min(xp)) | (q != np.max(xp))):
+    elif ((xp.dtype != int) | (1 != np.min(xp)) | (q != np.max(xp))):
         raise ValueError("The values of the existing design matrix x0 should be integers within: 1,2,3...,q.")
     elif (any(np.array([v for i in range(xp.shape[1]) for v in np.unique(xp[:, i], return_counts=True)[1]]) > n / q)):
         raise ValueError("xp does not follow a balanced design.")
@@ -553,7 +553,7 @@ def gen_aud_col(xp, n, s, q, init="rand", initX=np.array([[]]), crit="CD2", maxi
             raise ValueError("initX must be a numpy array.")
         elif ((n != initX.shape[0]) | (s <= initX.shape[1])):
             raise ValueError("The size of the input design matrix does not match the given n,s.")
-        elif ((1 != np.min(initX)) | (q != np.max(initX)) | (initX.dtype != np.int)):
+        elif ((1 != np.min(initX)) | (q != np.max(initX)) | (initX.dtype != int)):
             raise ValueError("The values of the input design matrix initX should be integers within: 1,2,3...,q.")
         elif (any(np.array([v for i in range(initX.shape[1]) for v in np.unique(initX[:, i], return_counts=True)[1]]) > n / q)):
             raise ValueError("initX does not follow a balanced design.")
@@ -744,7 +744,7 @@ def gen_aud_ms(xp, n, s, q, crit="CD2", maxiter=100, nshoot=5, random_state=0, n
         raise ValueError("xp must be a numpy array.")
     elif ((n <= xp.shape[0]) | (s != xp.shape[1])):
         raise ValueError("The size of the existing design matrix xp does not match the given n,s.")
-    elif ((xp.dtype != np.int) | (1 > np.min(xp)) | (q < np.max(xp))):
+    elif ((xp.dtype != int) | (1 > np.min(xp)) | (q < np.max(xp))):
         raise ValueError("The values of the existing design matrix x0 should be integers within: 1,2,3...,q.")
     elif (any(np.array([v for i in range(xp.shape[1]) for v in np.unique(xp[:, i], return_counts=True)[1]]) > n / q)):
         raise ValueError("xp does not follow a balanced design, please increase the number of n or remove duplicated elements (per column) in xp.")
@@ -855,7 +855,7 @@ def gen_aud_col_ms(xp, n, s, q, crit="CD2", maxiter=100, nshoot=5, random_state=
         raise ValueError("xp must be a numpy array.")
     elif ((n != xp.shape[0]) | (s <= xp.shape[1])):
         raise ValueError("The size of the existing design matrix xp does not match the given n,s.")
-    elif ((xp.dtype != np.int) | (1 != np.min(xp)) | (q != np.max(xp))):
+    elif ((xp.dtype != int) | (1 != np.min(xp)) | (q != np.max(xp))):
         raise ValueError("The values of the existing design matrix x0 should be integers within: 1,2,3...,q.")
     elif (any(np.array([v for i in range(xp.shape[1]) for v in np.unique(xp[:, i], return_counts=True)[1]]) > n / q)):
         raise ValueError("xp does not follow a balanced design.")
